(PC-17337)[PRO] refactor: remove duplicate and fix nav&logs bugs in Stock in Summary
