+++ pro/src/screens/OfferIndividual/Summary/StockSection/StockEventSection/StockEventItem/StockEventItem.tsx
