+++ pro/src/screens/OfferIndividual/Summary/StockSection/StockEventSection/StockEventItem/index.ts
