+++ pro/src/screens/OfferIndividual/Summary/StockSection/StockEventSection/index.ts
