+++ pro/src/screens/OfferIndividual/Summary/StockSection/StockSection.module.scss
