+++ pro/src/screens/OfferIndividual/Summary/StockSection/StockThingSection/index.ts
