+++ pro/src/screens/OfferIndividual/Summary/StockSection/index.ts
@@ -1,0 +1,1 @@
+export { default as StockSection } from './StockSection'
