+++ pro/src/screens/OfferIndividual/Summary/index.ts
@@ -1,5 +1,5 @@
 export { default as Summary } from './Summary'
 export type { ISummaryProps } from './Summary'
+export type { IStockEventItemProps } from './StockSection/StockEventSection'
+export type { IStockThingSectionProps } from './StockSection/StockThingSection'
 export type { IOfferSectionProps } from './OfferSection'
-export type { IStockThingSectionProps } from './StockThingSection'
-export type { IStockEventItemProps } from './StockEventSection'
